requirements.txt
python-telegram-bot==21.5
pylint==3.2.7